add missing part of DB's patch to convert flags and related variables into longs: file_flags should be declared as a long instead of an int
git-svn-id: svn://svn.savannah.gnu.org/nano/trunk/nano@1844 35c25a1d-7b9e-4130-9fde-d3aeb78583b8

+++ src/nano.h
@@ -160,7 +160,7 @@ typedef struct openfilestruct {
 				 * position. */
     int file_current_y;		/* Current file's y-coordinate
 				 * position. */
-    int file_flags;		/* Current file's flags: modification
+    long file_flags;		/* Current file's flags: modification
 				 * status (and marking status, if
 				 * available). */
     int file_placewewant;	/* Current file's place we want. */
